Refactor item link permission UI helpers
