+++ src/components/permissions/models/itemLinkPermissionModels.ts
@@ -1,6 +1,4 @@
 import type {
-  IItemLinkPermissionEntryForUI,
-  IItemUserPermissionRecipientForUI,
   ItemLinkPermissionRoleLabelForUI,
   ItemLinkPermissionScope,
   ItemLinkPermissionType,
@@ -87,51 +85,3 @@ export interface IItemLinkPermissionDerivedEntry {
   recipients: IItemLinkPermissionDisplayRecipient[];
   hasValidationError: boolean;
 }
-
-/**
- * 把 recipient 候选项还原成后端 `grant/revoke` 需要的最小合同。
- *
- * 优先级沿用现有 item permission 的 recipient 规则：
- * 优先 objectId，其次 email，最后 alias/UPN。
- *
- * @param candidate 前端当前持有的 recipient 候选项。
- * @returns 可直接提交给后端合同层的最小 recipient。
- */
-export const mapItemLinkRecipientCandidateToRequest = (
-  candidate: IItemLinkPermissionRecipientCandidate,
-): IItemUserPermissionRecipientForUI => ({
-  recipientObjectId: candidate.objectId,
-  recipientEmail: candidate.mail,
-  recipientAlias: candidate.userPrincipalName,
-});
-
-/**
- * 计算 links 面板是否存在本地未保存修改。
- *
- * @param draft 当前 links 草稿状态。
- * @returns 只要任一变更集合非空，就视为存在未保存修改。
- */
-export const hasItemLinkPermissionDraftChanges = (
-  draft: IItemLinkPermissionDraftState,
-): boolean =>
-  draft.createdLinks.length > 0 ||
-  draft.deletedPermissionIds.length > 0 ||
-  Object.keys(draft.grantsByPermissionId).length > 0 ||
-  Object.keys(draft.revokesByPermissionId).length > 0;
-
-/**
- * 生成 links 列表默认的空草稿状态。
- */
-export const createEmptyItemLinkPermissionDraftState =
-  (): IItemLinkPermissionDraftState => ({
-    createdLinks: [],
-    deletedPermissionIds: [],
-    grantsByPermissionId: {},
-    revokesByPermissionId: {},
-  });
-
-/**
- * 生成 links 面板默认的空后端快照。
- */
-export const createEmptyItemLinkPermissionEntries =
-  (): IItemLinkPermissionEntryForUI[] => [];
